refactor: extract conversation data handlers and delete redundant file
Move conversation data handlers (_now_iso_utc, _load_state,
handle_get/set/load_conversation_data) from sdk_tools.py into
huf/ai/handlers/conversation_data.py.

Delete huf/ai/conversation_data_tools.py which was a redundant copy
of the same functions already present in sdk_tools.py.

Backward-compatible re-exports kept in sdk_tools.py.
